Reorganizes some files
