+++ packages/sanes-browser-extension/src/routes/wallet/components/SidePanel/PanelDrawer/components/Requests/index.tsx
@@ -1,7 +1,7 @@
 import { Block, Typography } from "medulas-react-components";
 import React from "react";
 
-import { RequestContext } from "../../../../../../context/RequestProvider";
+import { RequestContext } from "../../../../../../../context/RequestProvider";
 import RequestList from "./RequestList";
 
 const Requests = (): JSX.Element => {
