+++ packages/sanes-browser-extension/src/routes/wallet/components/SidePanel/PanelDrawer/index.tsx
@@ -8,13 +8,13 @@ import chevronLeft from "../../../assets/chevronLeft.svg";
 import chevronRight from "../../../assets/chevronRight.svg";
 import closeButton from "../../../assets/closeButton.svg";
 import neumaLogo from "../../../assets/NeumaLogo.svg";
-import About from "./About";
-import DeleteWallet from "./DeleteWallet";
-import Menu from "./Menu";
-import Networks from "./Networks";
-import RecoveryWords from "./RecoveryWords";
-import Requests from "./Requests";
-import Settings from "./Settings";
+import About from "./components/About";
+import DeleteWallet from "./components/DeleteWallet";
+import Menu from "./components/Menu";
+import Networks from "./components/Networks";
+import RecoveryWords from "./components/RecoveryWords";
+import Requests from "./components/Requests";
+import Settings from "./components/Settings";
 
 export const drawerHtmlId = "wallet-sidepanel";
 export const toolbarHeight = 102;
